// src/auth/auth.model.d.ts
export interface claim {
    nombre: string;
    valor: string;
}

export interface credencialesUsuario{
    email: string;
    password: string;
}

export interface credencialesUsuarioValidator{
    email: string;
    password: string;
    passwordBackup: string;
}

export interface respuestaAutenticacion {
    token: string;
    expiracion: Date;
}

export interface usuarioDTO{
    id: string;
    email: string;
}